+++ index.php
@@ -6,6 +6,8 @@
  * @author   Maxim Ivanov <maxim_91@inbox.ru>
  */
 
+require('src/function.php');
+
 $name = 'Максим';
 $age = '27';
 
@@ -102,3 +104,21 @@ for ($tr = 1; $tr <= ROW; $tr++) {
 }
 $table .= '</table>';
 echo $table;
+
+
+
+$arrayStr = array('Привет', 'Андрей', 'как', 'дела');
+printString($arrayStr);
+
+calcEverything('*', 1, 2, 3, 4, 5);
+
+printTable(9, 8);
+echo '<br>';
+
+printDate();
+
+textEdit('Карл у Клары украл Кораллы');
+
+textReplace('Две бутылки лимонада', 'Две', 'Три');
+
+createTxt('text.txt');
